# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckoutSessionResponse": {
        "description": "Hosted Stripe Checkout URL for the founder unlock (DECISIONS.md D21).",
        "examples": [
          {
            "checkout_url": "https://checkout.stripe.com/c/pay/cs_test_a1b2c3",
            "session_id": "cs_test_a1b2c3"
          }
        ],
        "properties": {
          "checkout_url": {
            "description": "Open this URL to complete payment.",
            "title": "Checkout Url",
            "type": "string"
          },
          "session_id": {
            "description": "Stripe Checkout Session id.",
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "checkout_url",
          "session_id"
        ],
        "title": "CheckoutSessionResponse",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The body of an error response.",
        "properties": {
          "code": {
            "description": "Stable, machine-readable error code. Branch on this.",
            "examples": [
              "forbidden"
            ],
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured context. Never contains submitted values.",
            "title": "Details"
          },
          "message": {
            "description": "Human-readable explanation. May change; do not parse.",
            "examples": [
              "You do not have access to this resource."
            ],
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "description": "The one and only error response shape in this API.",
        "examples": [
          {
            "error": {
              "code": "forbidden",
              "message": "You do not have access to this resource."
            }
          },
          {
            "error": {
              "code": "validation_error",
              "details": {
                "fields": [
                  {
                    "field": "body.password",
                    "reason": "String should have at least 12 characters",
                    "type": "string_too_short"
                  }
                ]
              },
              "message": "The request failed validation."
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "UnlockReceiptResponse": {
        "description": "Completed unlock purchase, when one exists.",
        "examples": [
          {
            "amount": 14900,
            "currency": "USD",
            "paid_at": "2026-08-10T12:00:00Z",
            "reference": "3f2a1b0c-4d5e-6789-abcd-ef0123456789"
          }
        ],
        "properties": {
          "amount": {
            "description": "Amount in minor units (e.g. cents).",
            "title": "Amount",
            "type": "integer"
          },
          "currency": {
            "description": "ISO 4217 currency code.",
            "title": "Currency",
            "type": "string"
          },
          "paid_at": {
            "format": "date-time",
            "title": "Paid At",
            "type": "string"
          },
          "reference": {
            "description": "Our purchase id.",
            "title": "Reference",
            "type": "string"
          }
        },
        "required": [
          "reference",
          "amount",
          "currency",
          "paid_at"
        ],
        "title": "UnlockReceiptResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "description": "Access token issued by `POST /v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the FundReady platform. All resources live under `/v1`. Timestamps are ISO 8601 UTC, currencies are ISO 4217 codes, and money is expressed in integer minor units.\n\nEvery error response uses the envelope `{\"error\": {\"code\", \"message\", \"details\"}}`. Branch on `code`, which is stable; never parse `message`.\n\nEvery response carries an `X-Request-ID` header. Quote it in a support report and we can find the exact log line.",
    "summary": "AI business audit and investor-matching platform for SACI Holdings.",
    "title": "FundReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/billing/checkout": {
      "post": {
        "description": "Creates a Stripe Checkout Session for the one-off founder unlock (DECISIONS.md D21). Open `checkout_url` in a browser or system web view. Entitlement is granted only after Stripe delivers a verified `checkout.session.completed` webhook \u2014 do not treat opening the URL as payment.\n\n`422` with `already_unlocked` when `subscription_status` is already `active`.",
        "operationId": "start_checkout_v1_billing_checkout_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Start founder unlock Checkout",
        "tags": [
          "commerce"
        ]
      }
    },
    "/v1/billing/unlock": {
      "get": {
        "description": "Returns the founder's completed unlock purchase, if any.",
        "operationId": "read_unlock_v1_billing_unlock_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnlockReceiptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Completed unlock receipt",
        "tags": [
          "commerce"
        ]
      }
    },
    "/v1/billing/webhooks/stripe": {
      "post": {
        "description": "Verifies the Stripe signature header and applies entitlement changes. Idempotent on `event.id`. Not authenticated with a bearer token \u2014 Stripe signs the body instead.",
        "operationId": "stripe_webhook_v1_billing_webhooks_stripe_post",
        "parameters": [
          {
            "in": "header",
            "name": "Stripe-Signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stripe-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Stripe Webhook V1 Billing Webhooks Stripe Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "internal_error"
          }
        },
        "summary": "Stripe webhook receiver",
        "tags": [
          "commerce"
        ]
      }
    },
...
  }
}